feat(ci): upgrade write-addresses-json
Extracted common functionality for coming CI automation into module.
