+++ .github/scripts/package.json
@@ -1,4 +1,5 @@
 {
+  "type":"module",
   "dependencies": {
     "minimist": "^1.2.6"
   }
